Factor out adding deploy scripts for npm publish into a shared utility
