+++ packages/nx-plugin/src/generators/project/packageJson/index.ts
@@ -1,2 +1,3 @@
+export * from "./addNpmPublishDeployScriptsToPackageJson";
 export * from "./copyPackageJsonTemplate";
 export * from "./standardizePackageJson";
